Move json functionality into a separate file for dynamic message
